+++ src/components/events/SelectionChanged.ts
@@ -1,0 +1,10 @@
+/**
+ * Fired when document selection is changed
+ */
+export const SelectionChanged = 'selection changed';
+
+/**
+ * Payload that will be passed with the event
+ */
+export interface SelectionChangedPayload {
+}